refactor(io): extract PSD import/export into src/io/psd.ts
file-menu.ts was 565 lines because it doubled as the menu definition and
the implementation of every file-format operation. Moves the PSD code
(blend-mode tables, flattenLayerTreeForPsd, exportPsdFile, importPsdFile)
into src/io/psd.ts — format glue belongs with format code, not with menu
plumbing.

file-menu.ts is down to 189 lines and reads top-to-bottom as what it's
actually about: file-menu item definitions with thin handlers for open /
export. The two PSD exports are re-exported so App.tsx and any other
callers don't move.

The duplicated PSD BLEND_MODE tables still exist — the staff review
flagged these as a single-source-of-truth opportunity (codegen from
Rust). Leaving that for a follow-up PR since it requires a build-script.
